pyproject.toml
[project]
name = "cmugpt-agent"
version = "0.1.0"
description = "CMUGPT Agent for CMU campus information"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.0.0",
    "openai>=1.108.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = ["pre-commit>=4.2.0", "ruff>=0.11.10", "ty>=0.0.1a2"]

[tool.ty.terminal]
error-on-warning = true

[tool.ty.rules]
unused-ignore-comment = "warn"

[tool.ruff]
fix = true
unsafe-fixes = true

[tool.ruff.format]
docstring-code-format = true

# https://docs.astral.sh/ruff/linter/#rule-selection
[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
